roachtest: move getDiskUsageInBytes to own file
That way, when the tests are moved wholesale, this method
can move with them, simplifying the process without the
need to export this method in the interim.

Release note: None
